+++ app/src/main/java/com/jing/bilibilitv/http/data/LiveStreamUrlResponse.kt
@@ -7,9 +7,9 @@ data class LiveStreamUrlResponse(
     @SerializedName("accept_quality")
     val acceptQuality: List<String>,
     @SerializedName("current_qn")
-    val currentQn: Int,
+    val currentQn: Long,
     @SerializedName("current_quality")
-    val currentQuality: Int,
+    val currentQuality: Long,
     @SerializedName("durl")
     val durl: List<LiveStreamUrlDurl>,
     @SerializedName("quality_description")
@@ -18,13 +18,13 @@ data class LiveStreamUrlResponse(
 
 data class LiveStreamUrlDurl(
     @SerializedName("length")
-    val length: Int,
+    val length: Long,
     @SerializedName("order")
-    val order: Int,
+    val order: Long,
     @SerializedName("p2p_type")
-    val p2pType: Int,
+    val p2pType: Long,
     @SerializedName("stream_type")
-    val streamType: Int,
+    val streamType: Long,
     @SerializedName("url")
     val url: String
 )
@@ -33,5 +33,5 @@ data class LiveStreamUrlQualityDescription(
     @SerializedName("desc")
     val desc: String,
     @SerializedName("qn")
-    val qn: Int
+    val qn: Long
 )
